backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
python-multipart==0.0.6
openpyxl==3.0.9
pyarabic==0.6.14
googletrans==3.1.0a0
reportlab==3.6.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==0.7.0
jinja2==3.0.1
sqlalchemy==2.0.23
user-agents==2.2.0 